add sudo extension for extendGraphQl
