+++ .changeset/six-papayas-bob.md
@@ -1,0 +1,5 @@
+---
+'@keystone-6/core': minor
+---
+
+Adds a `sudo` parameter for `extendGraphqlSchema`, to indicate whether you are extending the sudo GraphQL schema or not
