+++ examples/extend-graphql-schema-graphql-ts/schema.graphql
@@ -13,6 +13,7 @@ type Post {
 enum PostStatusType {
   draft
   published
+  banned
 }
 
 scalar DateTime @specifiedBy(url: "https://datatracker.ietf.org/doc/html/rfc3339#section-5.6")
